Move `productInputAdditionsToSync`/``productInputDeletionsToSync` and dependent logic from `EditableOrderViewModel` to `ProductInputTransformer` - I'm unsure about this move after implementing this because these two functions aren't enough for the current POS UX, but committing the changes for now.
